+++ lib/resty/session/strategies/default.lua
@@ -4,12 +4,13 @@ local concat = table.concat
 local strategy = {}
 
 function strategy.load(session, cookie, key, keep_lock)
-  local storage = session.storage
-  local id      = cookie.id
+  local storage    = session.storage
+  local id         = cookie.id
+  local id_encoded = session.encoder.encode(id)
 
-  local data, err
+  local data, err, tag
   if storage.open then
-    data, err = session.storage:open(session.encoder.encode(id), keep_lock)
+    data, err = storage:open(id_encoded, keep_lock)
     if not data then
       return nil, err or "cookie data was not found"
     end
@@ -28,23 +29,50 @@ function strategy.load(session, cookie, key, keep_lock)
 
   local hkey = session.hmac(session.secret, key)
 
-  data, err = session.cipher:decrypt(data, hkey, id, session.key)
+  data, err, tag = session.cipher:decrypt(data, hkey, id, session.key, hash)
   if not data then
+    if storage.close then
+      storage:close(id_encoded)
+    end
+
     return nil, err or "unable to decrypt data"
   end
 
-  local input = concat{ key, data, session.key }
-  if session.hmac(hkey, input) ~= hash then
-    return nil, "cookie has invalid signature"
+  if tag then
+    if tag ~= hash then
+      if storage.close then
+        storage:close(id_encoded)
+      end
+
+      return nil, "cookie has invalid tag"
+    end
+
+  else
+    local input = concat{ key, data, session.key }
+    if session.hmac(hkey, input) ~= hash then
+      if storage.close then
+        storage:close(id_encoded)
+      end
+
+      return nil, "cookie has invalid signature"
+    end
   end
 
   data, err = session.compressor:decompress(data)
   if not data then
+    if storage.close then
+      storage:close(id_encoded)
+    end
+
     return nil, err or "unable to decompress data"
   end
 
   data, err = session.serializer.deserialize(data)
   if not data then
+    if storage.close then
+      storage:close(id_encoded)
+    end
+
     return nil, err or "unable to deserialize data"
   end
 
@@ -97,7 +125,6 @@ function strategy.modify(session, action, close, key)
     key = concat{ id, expires, usebefore }
   end
 
-  local hkey = session.hmac(session.secret, key)
   local data, err = session.serializer.serialize(session.data)
   if not data then
     if close and storage.close then
@@ -116,10 +143,11 @@ function strategy.modify(session, action, close, key)
     return nil, err or "unable to compress data"
   end
 
-  local hash = session.hmac(hkey, concat{ key, data, session.key })
+  local hkey = session.hmac(session.secret, key)
 
-  data, err = session.cipher:encrypt(data, hkey, id, session.key)
-  if not data then
+  local encrypted_data, tag
+  encrypted_data, err, tag = session.cipher:encrypt(data, hkey, id, session.key)
+  if not encrypted_data then
     if close and storage.close then
       storage:close(id_encoded)
     end
@@ -127,9 +155,18 @@ function strategy.modify(session, action, close, key)
     return nil, err
   end
 
+  local hash
+  if tag then
+    hash = tag
+  else
+    -- it would be better to calculate signature from encrypted_data,
+    -- but this is kept for backward compatibility
+    hash = session.hmac(hkey, concat{ key, data, session.key })
+  end
+
   if action == "save" and storage.save then
     local ok
-    ok, err = storage:save(id_encoded, ttl, data, close)
+    ok, err = storage:save(id_encoded, ttl, encrypted_data, close)
     if not ok then
       return nil, err
     end
@@ -151,8 +188,8 @@ function strategy.modify(session, action, close, key)
   if storage.save then
     cookie = concat({ id_encoded, expires, hash }, "|")
   else
-    data = session.encoder.encode(data)
-    cookie = concat({ id_encoded, expires, data, hash }, "|")
+    local encoded_data = session.encoder.encode(encrypted_data)
+    cookie = concat({ id_encoded, expires, encoded_data, hash }, "|")
   end
 
   return cookie
@@ -182,9 +219,11 @@ end
 
 function strategy.close(session)
   local id = session.id
-  local storage = session.storage
-  if id and storage.close then
-    return storage:close(session.encoder.encode(id))
+  if id then
+    local storage = session.storage
+    if storage.close then
+      return storage:close(session.encoder.encode(id))
+    end
   end
 
   return true