+++ Changes.md
@@ -2,16 +2,26 @@
 
 All notable changes to `lua-resty-session` will be documented in this file.
 
+## [3.9] - 2022-01-14
+### Fixed
+- Fix #138 issue of chunked cookies are not expired when session shrinks,
+  thanks @alexdowad.
+- Fix #134 where regenerate strategy destroyed previous session when calling
+  `session:regenerate`, it should just `ttl` the old session.
 
-## [3.9] - 2022-01-12
-### Changed
-- Optimize Redis and Memcache storage adapters to not connect to database
-  when not needed.
 ### Added
+- AES GCM mode support was added to AES cipher.
+  This is recommended, but for backward compatibility it was not set as default.
+  It will be changed in 4.0 release.
 - Redis ACL authentication is now available.
   - Add `session_redis_username`
   - Add `session_redis_password`
   - Deprecate `session_redis_auth`; use `session_redis_password`
+
+### Changed
+- Optimize Redis and Memcache storage adapters to not connect to database
+  when not needed.
+  
 
 ## [3.8] - 2021-01-04
 ### Added